Change sync hoc to a custom hook
