+++ src/ChangeAlert/useStorageListener.js
@@ -1,0 +1,21 @@
+import React from "react";
+
+function useStorageListener(sincronize) {
+  const [storageChange, setStorageChange] = React.useState(false);
+
+  // Con esto se subscribe a los cambios que se realicen en el local storage (y funciona en diferentes lugares que esten usando eso)
+  window.addEventListener("storage", (change) => {
+    if (change.key === "TODOS_V1") {
+      setStorageChange(true);
+    }
+  });
+
+  const toggleShow = () => {
+    sincronize();
+    setStorageChange(false);
+  };
+
+  return { show: storageChange, toggleShow };
+}
+
+export { useStorageListener };
